Refactor surface buffer size calculation.
This eliminates the duplication between pitchB() and size(), and
ensures that the latter is the full allocation size.

Bug chromium:835299

Reviewed-on: https://swiftshader-review.googlesource.com/18789
Tested-by: Nicolas Capens <nicolascapens@google.com>

+++ src/Renderer/Surface.hpp
@@ -362,7 +362,7 @@ namespace sw
 		static int pitchP(int width, int border, Format format, bool target);
 		static int sliceB(int width, int height, int border, Format format, bool target);
 		static int sliceP(int width, int height, int border, Format format, bool target);
-		static unsigned int size(int width, int height, int depth, int border, int samples, Format format);   // FIXME: slice * depth
+		static size_t size(int width, int height, int depth, int border, int samples, Format format);
 
 		static bool isStencil(Format format);
 		static bool isDepth(Format format);
